requirements.txt
# Database and ORM
sqlalchemy==1.4.46
psycopg2-binary==2.9.5
alembic==1.9.2
python-dotenv==0.21.0

# Web Framework
fastapi==0.88.0
uvicorn==0.20.0

# Data Processing
pandas==2.0.1
numpy==1.24.3

# Machine Learning
scikit-learn==1.2.2
xgboost==1.7.3

# Additional Utilities
pydantic==1.10.4
python-multipart==0.0.6

# Optional: for advanced data handling
pyarrow==11.0.0

# Logging
loguru==0.6.0